requirements.txt
notion-client>=2.0.0
schedule>=1.2.0
pyyaml>=6.0
markdown>=3.4.0
python-dotenv>=1.0.0
requests>=2.31.0
